+++ link-move/src/main/java/com/nhl/link/move/runtime/task/createorupdate/CreateOrUpdateSegment.java
@@ -8,7 +8,7 @@ import org.apache.cayenne.ObjectContext;
 /**
  * @since 1.3
  */
-public class CreateOrUpdateSegment implements DataSegment {
+public class CreateOrUpdateSegment extends DataSegment<CreateOrUpdateStage> {
 
     public static final String KEY_COLUMN = "$lm_key";
     public static final String TARGET_COLUMN = "$lm_target";
@@ -16,14 +16,6 @@ public class CreateOrUpdateSegment implements DataSegment {
 
     private final ObjectContext context;
     private final RowAttribute[] sourceRowsHeader;
-
-    private DataFrame sourceRows;
-    private DataFrame sources;
-    private DataFrame mappedSources;
-    private DataFrame matchedTargets;
-    private DataFrame mapped;
-    private DataFrame merged;
-    private DataFrame fksResolved;
 
     public CreateOrUpdateSegment(ObjectContext context, RowAttribute[] sourceRowsHeader) {
         this.context = context;
@@ -34,46 +26,46 @@ public class CreateOrUpdateSegment implements DataSegment {
         return context;
     }
 
+    public RowAttribute[] getSourceRowsHeader() {
+        return sourceRowsHeader;
+    }
+
     public DataFrame getSourceRows() {
-        return sourceRows;
+        return get(CreateOrUpdateStage.EXTRACT_SOURCE_ROWS);
     }
 
     /**
      * @since 2.17
      */
-    public CreateOrUpdateSegment setSourceRows(DataFrame sourceRows) {
-        this.sourceRows = sourceRows;
+    public CreateOrUpdateSegment setSourceRows(DataFrame df) {
+        set(CreateOrUpdateStage.EXTRACT_SOURCE_ROWS, df);
         return this;
     }
 
-    public RowAttribute[] getSourceRowsHeader() {
-        return sourceRowsHeader;
-    }
-
     public DataFrame getSources() {
-        return sources;
+        return get(CreateOrUpdateStage.CONVERT_SOURCE_ROWS);
     }
 
-    public CreateOrUpdateSegment setSources(DataFrame sources) {
-        this.sources = sources;
+    public CreateOrUpdateSegment setSources(DataFrame df) {
+        set(CreateOrUpdateStage.CONVERT_SOURCE_ROWS, df);
         return this;
     }
 
     public DataFrame getMappedSources() {
-        return mappedSources;
+        return get(CreateOrUpdateStage.MAP_SOURCE);
     }
 
-    public CreateOrUpdateSegment setMappedSources(DataFrame mappedSegment) {
-        this.mappedSources = mappedSegment;
+    public CreateOrUpdateSegment setMappedSources(DataFrame df) {
+        set(CreateOrUpdateStage.MAP_SOURCE, df);
         return this;
     }
 
     public DataFrame getMatchedTargets() {
-        return matchedTargets;
+        return get(CreateOrUpdateStage.MATCH_TARGET);
     }
 
-    public CreateOrUpdateSegment setMatchedTargets(DataFrame matchedTargets) {
-        this.matchedTargets = matchedTargets;
+    public CreateOrUpdateSegment setMatchedTargets(DataFrame df) {
+        set(CreateOrUpdateStage.MATCH_TARGET, df);
         return this;
     }
 
@@ -81,23 +73,14 @@ public class CreateOrUpdateSegment implements DataSegment {
      * @since 2.6
      */
     public DataFrame getMapped() {
-        return mapped;
+        return get(CreateOrUpdateStage.MAP_TARGET);
     }
 
     /**
      * @since 2.6
      */
-    public CreateOrUpdateSegment setMapped(DataFrame mapped) {
-        this.mapped = mapped;
-        return this;
-    }
-
-    public DataFrame getMerged() {
-        return merged;
-    }
-
-    public CreateOrUpdateSegment setMerged(DataFrame merged) {
-        this.merged = merged;
+    public CreateOrUpdateSegment setMapped(DataFrame df) {
+        set(CreateOrUpdateStage.MAP_TARGET, df);
         return this;
     }
 
@@ -105,14 +88,23 @@ public class CreateOrUpdateSegment implements DataSegment {
      * @since 2.12
      */
     public DataFrame getFksResolved() {
-        return fksResolved;
+        return get(CreateOrUpdateStage.RESOLVE_FK_VALUES);
     }
 
     /**
      * @since 2.12
      */
-    public CreateOrUpdateSegment setFksResolved(DataFrame fksResolved) {
-        this.fksResolved = fksResolved;
+    public CreateOrUpdateSegment setFksResolved(DataFrame df) {
+        set(CreateOrUpdateStage.RESOLVE_FK_VALUES, df);
+        return this;
+    }
+
+    public DataFrame getMerged() {
+        return get(CreateOrUpdateStage.MERGE_TARGET);
+    }
+
+    public CreateOrUpdateSegment setMerged(DataFrame df) {
+        set(CreateOrUpdateStage.MERGE_TARGET, df);
         return this;
     }
 }