+++ link-move/src/main/java/com/nhl/link/move/runtime/task/sourcekeys/SourceKeysSegment.java
@@ -1,18 +1,15 @@
 package com.nhl.link.move.runtime.task.sourcekeys;
 
 import com.nhl.link.move.RowAttribute;
-import org.dflib.DataFrame;
 import com.nhl.link.move.runtime.task.common.DataSegment;
+import org.dflib.DataFrame;
 
 /**
  * @since 1.3
  */
-public class SourceKeysSegment implements DataSegment {
+public class SourceKeysSegment extends DataSegment<SourceKeysStage> {
 
     private final RowAttribute[] sourceRowsHeader;
-
-    private DataFrame sourceRows;
-    private DataFrame sources;
 
     public SourceKeysSegment(RowAttribute[] sourceRowsHeader) {
         this.sourceRowsHeader = sourceRowsHeader;
@@ -23,23 +20,23 @@ public class SourceKeysSegment implements DataSegment {
     }
 
     public DataFrame getSourceRows() {
-        return sourceRows;
+        return get(SourceKeysStage.EXTRACT_SOURCE_ROWS);
     }
 
     /**
      * @since 3.0.0
      */
-    public SourceKeysSegment setSourceRows(DataFrame sourceRows) {
-        this.sourceRows = sourceRows;
+    public SourceKeysSegment setSourceRows(DataFrame df) {
+        set(SourceKeysStage.EXTRACT_SOURCE_ROWS, df);
         return this;
     }
 
     public DataFrame getSources() {
-        return sources;
+        return get(SourceKeysStage.CONVERT_SOURCE_ROWS);
     }
 
-    public SourceKeysSegment setSources(DataFrame sources) {
-        this.sources = sources;
+    public SourceKeysSegment setSources(DataFrame df) {
+        set(SourceKeysStage.CONVERT_SOURCE_ROWS, df);
         return this;
     }
 }